Torch added, compatibility python
