+++ GPREPS_gym/benchmarks.py
@@ -7,7 +7,7 @@ def bench(env, hipol, pol, verbose = False):
     '''
     N = 100
     R = np.zeros(N)
-    for rollout in xrange(N):
+    for rollout in range(N):
         x = env.reset()                    # Sample context
         w = hipol.mean(x.reshape(1, -1)).T # Sample lower-policy weights
